extract amount selector to component
