+++ tsconfig.json
@@ -27,7 +27,8 @@
       "@stores/*": ["app/stores/*"],
       "@utils/*": ["app/utils/*"],
       "@core/*": ["app/core/*"],
-      "@ui-library/*": ["app/ui-library/*"]
+      "@ui-library/*": ["app/ui-library/*"],
+      "@ui-components/*": ["app/ui-components/*"]
     }
   },
   "include": ["src", "styled.d.ts"]
